% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{f16qamdemod}
\alias{f16qamdemod}
\title{f16qamdemod}
\usage{
f16qamdemod(r)
}
\arguments{
\item{bits:}{vector of bits}
}
\description{
Receives a vector of complex values, r, corresponding to a
16-QAM modulated signal transmitted through a communications channel
(e.g., signal plus noise). Each received 16-QAM symbol (one symbol per r sample) is decoded into
4 bits. The 16-QAM symbol decision regions are defined with respect to
the constellation generated by f16qammod, \eqn{E_{s}}{Es} = 10, where in-phase and quadrature
constellation points take on values of (-3, -1, +1, +3), respectively.
}

