% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{f64qamdemod}
\alias{f64qamdemod}
\title{f64qamdemod}
\usage{
f64qamdemod(r)
}
\arguments{
\item{bits:}{vector of bits}
}
\description{
Receives a vector of complex values, r, corresponding to a
64-QAM modulated signal transmitted through a communications channel
(e.g., signal plus noise). Each received 64-QAM symbol (one symbol per r sample) is decoded into
6 bits. The 64-QAM symbol decision regions are defined with respect to
the constellation generated by f64qammod, \eqn{E_{s}}{Es} = 42, where in-phase and quadrature
constellation points take on values of (-7, -5, -3, -1, +1, +3, +5, +7), respectively.
}

