% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fbpskdemod}
\alias{fbpskdemod}
\title{fbpskdemod}
\usage{
fbpskdemod(r)
}
\arguments{
\item{r:}{received signal plus noise.}
}
\description{
Receives a vector of real values, corresponding to a
BPSK modulated signal transmitted through a communications channel
(e.g., signal plus noise). An input value < 1 is mapped to an
output value of 0, otherwise to a value of 1.
}

