% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fbpskmod}
\alias{fbpskmod}
\title{fbpskmod}
\usage{
fbpskmod(bits)
}
\arguments{
\item{bits:}{received vector of bits (0's and 1's).}
}
\description{
BPSK modulator receives a vector of bits, each with value 0 or 1, and outputs a
vector with values 1 and -1. An output value of -1 corresponds to an input value
of 0 and an output value of 1 corresponds to an input of 1.
}

