% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fNo}
\alias{fNo}
\title{fNo function}
\usage{
fNo(N, No, Type = "real")
}
\arguments{
\item{N}{number of noise samples}

\item{No:}{noise spectral density}

\item{Type:}{"real" or "complex" defualts to real}
}
\description{
Generate a vector of normally distributed noise samples with mean of zero and noise spectral density (No/2), a.k.a. AWGN.
}

