% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fr8pskbitmap}
\alias{fr8pskbitmap}
\title{f8pskdemod}
\usage{
fr8pskbitmap(r)
}
\arguments{
\item{r:}{received signal.}
}
\description{
Receives a vector of complex values, r, corresponding to an
8-PSK modulated signal transmitted through a communications channel
(e.g., signal plus noise). Three bits are output for each received symbol
according to the following decision rules
\tabular{cc}{
input \tab output \cr
 \eqn{ -\pi/8 \gee Arg(r) < \pi/8} \tab 000  \cr
 \eqn{  \pi/8 \ge Arg(r) < 3 \pi/8} \tab 001 \cr
 \eqn{ 3 \pi/8 \ge Arg(r) < 5 \pi/8} \tab 011 \cr
 \eqn{ 5 \pi/8 \ge Arg(r) < 7 \pi/8} \tab 010 \cr
 \eqn{ 7 \pi/8 \ge Arg(r) < 9 \pi/8} \tab 110 \cr
 \eqn{ -7 \pi/8 \ge Arg(r) < -5 \pi/8} \tab 111 \cr
 \eqn{ -5 \pi/8 \ge Arg(r) < -3 \pi/8} \tab 101 \cr
 \eqn{ -3 \pi/8 \ge Arg(r) < - \pi/8}\tab  101 \cr
}
}

