% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{iqdensityplot}
\alias{iqdensityplot}
\title{IQ Density Plot}
\usage{
iqdensityplot(r, iq = "r")
}
\arguments{
\item{r}{- complex or real valued vector}

\item{iq}{- if iq = "r" (default) then plot density of Re(r) else if iq = "q" then plot density of Im(r)}
}
\description{
A convenience function to plot a density function of a vector containing the in-phase and
quadrature signal (plus noise).
}
\examples{
M=4
Es=1
Eb = Es/log2(M)
Nsymbols=1000
Nbits=log2(M)*Nsymbols
bits <- sample(0:1,Nbits, replace=TRUE)
s <- fqpskmod(bits)
EbNodB=4
No = Eb/(10^(EbNodB/10))
n <- fNo(Nsymbols,No,type="complex")
r <- s+n
 @family rwirelesscom functions
}

