% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{iqscatterplot}
\alias{iqscatterplot}
\title{iqscatterplot}
\usage{
iqscatterplot(r)
}
\arguments{
\item{r}{complex or real valued vector}
}
\description{
A convenience function that plots a scatter diagram of Im(r) vs. Re(r). The function is
useful for visualing constellations such as M-PSK or M-QAM.
}
\seealso{
Other modulation.demodulation.communications: \code{\link{f16qamdemod}};
  \code{\link{f16qammod}}; \code{\link{f64qamdemod}};
  \code{\link{f64qammod}}; \code{\link{f8pskdemod}};
  \code{\link{f8pskmod}}; \code{\link{fNo}};
  \code{\link{fbpskdemod}}; \code{\link{fbpskmod}};
  \code{\link{fqpskdemod}}; \code{\link{fqpskmod}};
  \code{\link{iqdensityplot}}
}

