% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{stemplot}
\alias{stemplot}
\title{Stem Plot}
\usage{
stemplot(x, y, pch = 16, linecol = 1, clinecol = 1, linew = 1, ...)
}
\arguments{
\item{x}{- vector of points}
}
\description{
afsadf Receives a vector of points and plots a stemplot
}
\examples{
x <- seq(-8*pi, 8*pi, by = 0.1)
y <- sinc(x)
stemplot(x/pi,y,ylim=c(-0.3,1.1), pch=19, cex=0.2, ylab="", xlab="")
}

