% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fbpskmod}
\alias{fbpskmod}
\title{BPSK Modulator}
\usage{
fbpskmod(bits)
}
\arguments{
\item{bits}{- vector of bits (0's and 1's)}
}
\value{
returns a BPSK modulated vector, each element taking on values of 1 or -1
}
\description{
Receives a vector of bits, each with value 0 or 1, and outputs a
vector with values 1 and -1, respectively.
}
\examples{
bits <- sample(0:1,10, replace=TRUE)
fbpskmod(bits)
}
\seealso{
Other modulation.demodulation.communications: \code{\link{f16pskdemod}};
  \code{\link{f16pskmod}}; \code{\link{f16qamdemod}};
  \code{\link{f16qammod}}; \code{\link{f64qamdemod}};
  \code{\link{f64qammod}}; \code{\link{f8pskdemod}};
  \code{\link{f8pskmod}}; \code{\link{fNo}};
  \code{\link{fbpskdemod}}; \code{\link{fqpskdemod}};
  \code{\link{fqpskmod}}; \code{\link{iqdensityplot}};
  \code{\link{iqscatterplot}}
}

