% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fqpskdemod}
\alias{fqpskdemod}
\title{fqpskdemod}
\usage{
fqpskdemod(r)
}
\arguments{
\item{r:}{received signal plus noise.}
}
\description{
Receives a vector of complex values, r, corresponding to a
QPSK modulated signal transmitted through a communications channel
(e.g., signal plus noise). The received signal, r, is mapped to its in-phase (real part) and
quadrature parts (imaginary part) and demodulated, such that two binary bits are output
for each value of r. If the in phase part is > 0 then the corresponding output bit value is 1,
otherwise 0. Similarly, if the quadrature part (imaginary) > 0 then the corresponding bit value
is 1, otherwise 0.
}
\seealso{
Other modulation.demodulation.communications: \code{\link{f16qamdemod}};
  \code{\link{f16qammod}}; \code{\link{f64qamdemod}};
  \code{\link{f64qammod}}; \code{\link{f8pskdemod}};
  \code{\link{f8pskmod}}; \code{\link{fNo}};
  \code{\link{fbpskdemod}}; \code{\link{fbpskmod}};
  \code{\link{fqpskmod}}; \code{\link{iqdensityplot}};
  \code{\link{iqscatterplot}}
}

