% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{fqpskmod}
\alias{fqpskmod}
\title{fqpskmod}
\usage{
fqpskmod(bits)
}
\arguments{
\item{bits:}{received vector of bits (0's and 1's).}
}
\description{
Receives a vector of bits (1's and 0's). The 1's and 0's are
mapped to in-phase (real) and quadrature (imaginary) components.
Correspondingly, a bit of 1 is mapped to +1/sqrt(2), otherwise to -1/sqrt(2)
as follows:
 \tabular{cc}{
input \tab output \cr
00 \tab  (-1 - 1j) / sqrt(2)  \cr
01 \tab  (-1 + 1j) / sqrt(2) \cr
10 \tab  (+1 - 1j) / sqrt(2) \cr
11 \tab  (+1 + 1j) / sqrt(2)
}
}

