% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{iqdensityplot}
\alias{iqdensityplot}
\title{iqdensityplot}
\usage{
iqdensityplot(r, iq = "r")
}
\arguments{
\item{r:}{complex or real valued vector}

\item{iq:}{if = "r" (default) then plot desnity of Re(r) else if iq = "i" the plot density of Im(r)}
}
\description{
A convenidnece function to plot a desnity function of a vector containing the in-phase and
quadrature signal (plus noise).
}

