% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rwirelesscom.R
\name{iqscatterplot}
\alias{iqscatterplot}
\title{iqscatterplot}
\usage{
iqscatterplot(r)
}
\arguments{
\item{r:}{complex or real valued vector}
}
\description{
A convenience function that plots a scatter diagram of Im(r) vs. Re(r). The function is
useful for visualing constellations such as M-PSK or M-QAM.
}

